Created dedicated respawn tab
Created a dedicated repsawn tab to keep the code fairly clean and simple.
