+++ main.js
@@ -1,59 +1,32 @@
-
-var roleHarvester = require('role.harvester');
+//TODO: setup harvesters so that one massive harvester collects and carrys just coem grab the resources from it.
+//TODO: Put the creeper count values into variables that can be easily set from a file.
+//TODO: Look into resource search distance, creepers are going to the wrong side of the minieral where they can't even harvest because it is too far into the rock.
+//TODO: Create AND checks for creeper spawns so that it makes sure the needed creepers get spawned before the not so much needed ones.
 var roleBuilder = require('role.builder');
+var roleHarvester = require('role.harvester');
 var roleUpgrader = require('role.upgrader');
-
-
-
-var builderTotal = 3;
-var harvesterTotal = 2;
-var upgraderTotal = 2;
-
+var roleRespawner = require('role.respawner');
 
 module.exports.loop = function () {
     
-    
-    var harvesters = _.filter(Game.creeps, (creep) => creep.memory.role == 'harvester');
-    var builders = _.filter(Game.creeps, (creep) => creep.memory.role == 'builder');
-    var upgraders = _.filter(Game.creeps, (creep) => creep.memory.role == 'upgrader');
-    //console.log('Harvesters: ' + harvesters.length);
+    roleRespawner.run();
 
-    if(harvesters.length < harvesterTotal) {
-        var newName = 'Harvester' + Game.time;
-        console.log('Spawning new harvester: ' + newName);
-        Game.spawns['Spawn1'].spawnCreep([WORK,CARRY,MOVE], newName,
-            {memory: {role: 'harvester'}});
-    }
-    if(builders.length < builderTotal) {
-        var newName = 'Builder' + Game.time;
-        console.log('Spawning new builder: ' + newName);
-        Game.spawns['Spawn1'].spawnCreep([WORK,CARRY,MOVE], newName,
-            {memory: {role: 'builder'}});
-    }
-    if(upgraders.length < upgraderTotal) {
-        var newName = 'Upgrader' + Game.time;
-        console.log('Spawning new upgrader: ' + newName);
-        Game.spawns['Spawn1'].spawnCreep([WORK,CARRY,MOVE], newName,
-            {memory: {role: 'upgrader'}});
-    }
 
 
     for(var name in Game.creeps) {
         var creep = Game.creeps[name];
         
+        if(creep.memory.role == 'harvester')
         {
             roleHarvester.run(creep);
         }
+        
         if(creep.memory.role == 'builder'){
             roleBuilder.run(creep);
         }
+        
         if(creep.memory.role == 'upgrader'){
             roleUpgrader.run(creep);
         }
-        if(creep.memory.role == 'harvester'){
-            roleHarvester.run(creep);
-        }
-        
-        
     }
 }
